+++ lib/src/contracts/printable_widget.dart
@@ -1,6 +1,0 @@
-import 'package:flutter/material.dart';
-
-abstract class PrintableWidget {
-  void printWidget(BuildContext context);
-  String? toRealStringShallow(BuildContext context);
-}